Fixed pump short cycle
Fixed issue with pump coming on for one second.  The problem was the
heat would be needed and pump would be on, but pump amps would not be
high enough to turn on heater.  Also shortened minimum heater on time
from 3 min to 1 min.  There seems to be an issue with Arduino Mega
locking up now.